chore: use validation pkg
