+++ packages/validation/src/errors.ts
@@ -9,3 +9,24 @@ export class DRPValidationError extends Error {
 		this.name = "DRPValidationError";
 	}
 }
+
+export class InvalidHashError extends Error {
+	constructor(message: string = "Invalid hash") {
+		super(message);
+		this.name = "InvalidHashError";
+	}
+}
+
+export class InvalidDependenciesError extends Error {
+	constructor(message: string = "Invalid dependencies") {
+		super(message);
+		this.name = "InvalidDependenciesError";
+	}
+}
+
+export class InvalidTimestampError extends Error {
+	constructor(message: string = "Invalid timestamp") {
+		super(message);
+		this.name = "InvalidTimestampError";
+	}
+}
