+++ packages/validation/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./schemas/object.js";
+export * from "./vertex.js";
