+++ packages/validation/src/vertex.ts
@@ -1,34 +1,36 @@
 import { type IHashGraph, type Vertex } from "@ts-drp/types";
 import { computeHash } from "@ts-drp/utils/hash";
 
+import { InvalidDependenciesError, InvalidHashError, InvalidTimestampError } from "./errors.js";
+
 export interface ValidationResult {
 	success: boolean;
 	error?: Error;
 }
 
-function validateVertexHash(vertex: Vertex): void {
-	const correctHash = computeHash(vertex.peerId, vertex.operation, vertex.dependencies, vertex.timestamp);
-	if (vertex.hash !== correctHash) {
-		throw new Error(`Vertex ${vertex.hash} has invalid hash`);
+function validateVertexHash({ hash, peerId, operation, dependencies, timestamp }: Vertex): void {
+	const correctHash = computeHash(peerId, operation, dependencies, timestamp);
+	if (hash !== correctHash) {
+		throw new InvalidHashError(`Invalid hash for vertex ${hash}`);
 	}
 }
 
-function validateVertexDependencies(vertex: Vertex, hashGraph: IHashGraph): void {
-	if (vertex.dependencies.length === 0) {
-		throw new Error(`Vertex ${vertex.hash} has no dependencies`);
+function validateVertexDependencies({ hash, dependencies, timestamp }: Vertex, hashGraph: IHashGraph): void {
+	if (dependencies.length === 0) {
+		throw new InvalidDependenciesError(`Vertex ${hash} has no dependencies.`);
 	}
-	for (const dep of vertex.dependencies) {
+	for (const dep of dependencies) {
 		const depVertex = hashGraph.vertices.get(dep);
 		if (depVertex === undefined) {
-			throw new Error(`Vertex ${vertex.hash} has invalid dependency ${dep}`);
+			throw new InvalidDependenciesError(`Vertex ${hash} has invalid dependency ${dep}.`);
 		}
-		validateVertexTimestamp(depVertex.timestamp, vertex.timestamp, vertex.hash);
+		validateVertexTimestamp(depVertex.timestamp, timestamp, hash);
 	}
 }
 
 function validateVertexTimestamp(a: number, b: number, hash: string): void {
 	if (a > b) {
-		throw new Error(`Vertex ${hash} has invalid timestamp`);
+		throw new InvalidTimestampError(`Vertex ${hash} has invalid timestamp ${a} > ${b}`);
 	}
 }
